+++ src/modules/dashboard/users/components/layouts/Dashboardfooter.jsx
@@ -1,0 +1,14 @@
+import React from 'react'
+
+const Dashboardfooter = () => {
+  return (
+    <>
+       {/* Footer copyright */}
+       <article className='py-4 px-2 bg-gray-700 text-white absolute bottom-0 left-0 right-0'>
+          <p className='text-center'>Copyright&copy;2023Elimuendelezi.co.ke</p>
+        </article>
+    </>
+  )
+}
+
+export default Dashboardfooter
