+++ src/routes/routes.jsx
@@ -5,7 +5,7 @@ import {
   } from "react-router-dom";
   import App from "../App";
 
-  import { Error, Home, Courses, Blog, AboutUs, ContactUs, SingleBlog, SingleCourse, Login, Register, Terms, Dashboard, UserProfile, ChangePassword, Courseonoffer, Blogpage,Generatereport, AdminDashoboard, } from "../views";
+  import { Error, Home, Courses, Blog, AboutUs, ContactUs, SingleBlog, SingleCourse, Login, Register, Terms, Dashboard, DashboardHome, UserProfile, ChangePassword, Courseonoffer, Blogpage,Generatereport, AdminDashoboard, } from "../views";
   import { loginAction } from "../auth/login/services/loginAction";
   import { registerAction } from "../auth/signup/services/registerAction";
   const routes = createBrowserRouter(
@@ -32,12 +32,14 @@ import {
           <Route element={<Terms />} path="terms" />
 
           {/* Portal */}
-          <Route element={<Dashboard />} path="dashboard" />
-          <Route element={<UserProfile />} path="profile" />
-          <Route element={<ChangePassword />} path="changepassword" />
-          <Route element={<Courseonoffer />} path="coursesonoffer" />
-          <Route element={<Blogpage />} path="blogpage" />
-          <Route element={<Generatereport />} path="downloadcoursesummary" />
+          <Route element={<Dashboard />} path="dashboard">
+            <Route element={<DashboardHome />} index />
+            <Route element={<UserProfile />} path="profile" />
+            <Route element={<ChangePassword />} path="changepassword" />
+            <Route element={<Courseonoffer />} path="coursesonoffer" />
+            <Route element={<Blogpage />} path="blogpage" />
+            <Route element={<Generatereport />} path="downloadcoursesummary" />
+          </Route>
 
           <Route 
             element={<AdminDashoboard />}
