+++ src/views/index.js
@@ -9,13 +9,13 @@ import ContactUs from '../modules/contacts_and_about_us/view/ContactUs'
 import SingleBlog from '../modules/blog/views/SingleBlog'
 import SingleCourse from '../modules/courses/view/SingleCourse'
 import Terms from '../modules/terms_and_policies/views/Terms'
-import Dashboard from '../modules/dashboard/view/Dashboard'
-import UserProfile from '../modules/dashboard/view/UserProfile'
-import ChangePassword from '../modules/dashboard/view/ChangePassword'
-import Courseonoffer from '../modules/dashboard/view/Courseonoffer'
-import Blogpage from '../modules/dashboard/view/Blogpage'
-import Generatereport from '../modules/dashboard/view/Genratereport'
+import UserProfile from '../modules/dashboard/users/views/UserProfile'
+import ChangePassword from '../modules/dashboard/users/views/ChangePassword'
+import Courseonoffer from '../modules/dashboard/users/views/Courseonoffer'
+import Blogpage from '../modules/dashboard/users/views/Blogpage'
+import Generatereport from '../modules/dashboard/users/views/Genratereport'
 import Dashboard from '../modules/dashboard/users/views/Dashboard'
+import DashboardHome from "../modules/dashboard/users/views/DashboardHome"
 import AdminDashoboard from '../modules/dashboard/admin/views/AdminDashoboard'
 
 export {
@@ -37,5 +37,6 @@ export {
     Blogpage,
     Generatereport,
     AdminDashoboard,
+    DashboardHome,
 }
